requirements.txt
mlflow
numpy
notebook
pandas
python-box
scikit-learn
scipy
seaborn
matplotlib
pyYAML
types-pyYAML
ensure
tqdm
joblib